+++ backend/src/MonChenil.Api/Controllers/ReservationsController.cs
@@ -1,6 +1,7 @@
 using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using MonChenil.Api.Requests;
 using MonChenil.Domain.Pets;
 using MonChenil.Domain.Reservations;
 
@@ -28,18 +29,17 @@ public class ReservationsController : ControllerBase
     [HttpGet]
     public IActionResult GetCurrentUserReservations()
     {
-        string ownerId = GetCurrentUserId();
-        return Ok(reservationsRepository.GetReservations().Where(reservation => reservation.OwnerId == ownerId));
+        string currentUserId = GetCurrentUserId();
+        return Ok(reservationsRepository.GetReservationsByOwnerId(currentUserId));
     }
 
     [HttpPost]
-    public IActionResult CreateReservation(ReservationDto reservationDto)
+    public IActionResult CreateReservation(CreateReservationRequest request)
     {
-        string ownerId = GetCurrentUserId();
+        string currentUserId = GetCurrentUserId();
         var reservationId = new ReservationId(Guid.NewGuid());
-        var reservation = new Reservation(reservationId, ownerId, reservationDto.StartDate, reservationDto.EndDate);
-
-        List<Pet> pets = petsRepository.GetPets().Where(pet => reservationDto.PetIds.Contains(pet.Id)).ToList();
+        var reservation = new Reservation(reservationId, currentUserId, request.StartDate, request.EndDate);
+        var pets = petsRepository.GetPetsByIds(request.PetIds);
         reservation.AddPets(pets);
 
         reservationsRepository.AddReservation(reservation);
@@ -49,12 +49,17 @@ public class ReservationsController : ControllerBase
     [HttpDelete("{id:guid}")]
     public IActionResult DeleteReservation(Guid id)
     {
-        string ownerId = GetCurrentUserId();
-        ReservationId reservationId = new ReservationId(id);
-        var reservation = reservationsRepository.GetReservations().FirstOrDefault(reservation => reservation.Id == reservationId && reservation.OwnerId == ownerId);
+        string currentUserId = GetCurrentUserId();
+        ReservationId reservationId = new(id);
+        var reservation = reservationsRepository.GetReservationById(reservationId);
         if (reservation == null)
         {
             return NotFound();
+        }
+
+        if (reservation.OwnerId != currentUserId)
+        {
+            return Forbid();
         }
 
         reservationsRepository.DeleteReservation(reservation);
